Trying a new approach
A Logger class was simplified, it behaves like a logger, with
info, warn, error, and debug methods.
The LoggerRequest class was also simplified, the express-winston module
was removed because it was cluttering our interface too much.
An example app was created under the tests/ folder, so anyone can try
the changes more easily.